chore(server): move session history toolset module
